Adding `ArrayOrValue` to properties on EntityType
