+++ packages/graph/hash_graph/tests/test_data/entity_type/book.json
@@ -5,7 +5,10 @@
   "type": "object",
   "properties": {
     "https://blockprotocol.org/@alice/types/property-type/name": {
-      "$ref": "https://blockprotocol.org/@alice/types/property-type/name/v/1"
+      "type": "array",
+      "items": {
+        "$ref": "https://blockprotocol.org/@alice/types/property-type/name/v/1"
+      }
     },
     "https://blockprotocol.org/@alice/types/property-type/blurb": {
       "$ref": "https://blockprotocol.org/@alice/types/property-type/blurb/v/1"
